fix(training): Resolve multiple PyTorch training compatibility issues
- Fix tensor dtype mismatch in processor.py by explicitly setting
  dtype=torch.long for text_token and instruct_token tensors
- Fix ProcessGroup.options AttributeError in train_utils.py for
  newer PyTorch versions by using fixed 60-second timeout
- Fix tensor dimension mismatch in flow.py by adding length alignment
  for h and feat tensors that can differ by 1 due to rounding
- Add Stage 5 (LLM) completion detection in run_all_stages.sh to
  skip already-completed training epochs

These fixes enable successful Czech language fine-tuning on
Fun-CosyVoice3-0.5B-2512 model.

+++ cosyvoice/dataset/processor.py
@@ -387,10 +387,10 @@ def padding(data, use_spk_embedding, mode='train', gan=False, dpo=False):
                                    batch_first=True,
                                    padding_value=0)
         text = [sample[i]['text'] for i in order]
-        text_token = [torch.tensor(sample[i]['text_token']) for i in order]
+        text_token = [torch.tensor(sample[i]['text_token'], dtype=torch.long) for i in order]
         text_token_len = torch.tensor([i.size(0) for i in text_token], dtype=torch.int32)
         text_token = pad_sequence(text_token, batch_first=True, padding_value=0)
-        instruct_token = [torch.tensor(sample[i]['instruct_token']) for i in order]
+        instruct_token = [torch.tensor(sample[i]['instruct_token'], dtype=torch.long) for i in order]
         instruct_token_len = torch.tensor([i.size(0) for i in instruct_token], dtype=torch.int32)
         instruct_token = pad_sequence(instruct_token, batch_first=True, padding_value=0)
         utt_embedding = torch.stack([sample[i]['utt_embedding'] for i in order], dim=0)
